requirements.txt
Django==1.11.6
djangorestframework==3.7.1
uWSGI==2.0.15
django-cors-headers==2.1.0
django-jet==1.0.6
jsonfield==2.0.2
python-telegram-bot==8.1.1
